Finition vue and CAD
Implementation of the view with graph
Implementation of communication with arduino and simulation of data

+++ src/exia/fr/ModelObserver.java
@@ -1,11 +1,11 @@
 package exia.fr;
 
 public interface ModelObserver {
-	void onTemperatureIntChanged(double value);
+	void onTemperatureIntChanged(float value);
 
-	void onHumidityChanged(double value);
+	void onHumidityChanged(float value);
 
-	void onTemperatureConsigneChanged(double value);
+	void onTemperatureConsigneChanged(float value);
 
 	void onAlertChange(int value);
 }
